feat: make zoomed out elements centered and fully visible
The zoomable element should be always visible after initialization, hence
we have to scale either to the width or height, whatever results in the
smaller zoom-factor. However, in order to get the accurate zoom factor for
the y-dimension we need to set the aspect ratio depending on the parent
dimensions.

Also, the element should always be centered, if it doesn't have the same
dimensions as the container element. This includes the feature that it
always returns to the centered position upon zooming out.

If the window is resized, e.g. because the user rotates their phone from
portrait to landscape, than the initial offset needs to be recalculated
to ensure that the element is still centered and fully visible when
zooming out.

If the pictures are already in the browsers cache, the load event may be
fired before this zoom library is attached. Hence this ensures that the
code to center the element always runs at least once upon initialization.

When using this library for cases in which the images are continually
replaced, then always resetting the zoom and offset is not desirable.

We cannot simply handle just the first load event, because cached images may
not cause any, or the library may be attached after the event fired.

Hence we choose to wait 2 seconds for any load event and afterwards only
adjust the initial offset silently.